requirements.txt
scikit-learn == 1.0.1
scikit-image >= 0.19.2
pyradiomics >= 3.0.1
simpleitk >= 2.1.1
scipy >= 1.7.1
pandas >= 1.1.4
pingouin >= 0.4.0
tqdm >= 4.48.2
mri-normalization-tools >= 0.1
seaborn
RENT
torchio
